change package pics to media. rename class Image to Media.
