+++ twitter4j-core/src/main/java/twitter4j/media/impl/TwitgooOAuthUploader.java
@@ -24,20 +24,20 @@ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-package twitter4j.pics.impl;
+package twitter4j.media.impl;
 
 import twitter4j.TwitterException;
 import twitter4j.http.OAuthAuthorization;
 import twitter4j.internal.http.HttpParameter;
-import twitter4j.pics.AbstractImageUploader;
-import twitter4j.pics.ImageUploadException;
+import twitter4j.media.AbstractMediaUploader;
+import twitter4j.media.MediaUploadException;
 
 /**
  * @author Takao Nakaguchi - takao.nakaguchi at gmail.com
  * @author withgod - noname at withgod.jp
  * @since Twitter4J 2.1.8
  */
-public class TwitgooOAuthUploader extends AbstractImageUploader {
+public class TwitgooOAuthUploader extends AbstractMediaUploader {
 
     public TwitgooOAuthUploader(OAuthAuthorization oauth) {
         super(oauth);
@@ -45,7 +45,7 @@ public class TwitgooOAuthUploader extends AbstractImageUploader {
 
 
     @Override
-    public String postUp() throws TwitterException, ImageUploadException {
+    public String postUp() throws TwitterException, MediaUploadException {
         int statusCode = httpResponse.getStatusCode ();
         if (statusCode != 200)
             throw new TwitterException ("Twitgoo image upload returned invalid status code", httpResponse);
@@ -76,7 +76,7 @@ public class TwitgooOAuthUploader extends AbstractImageUploader {
     }
 
     @Override
-    public void preUp() throws TwitterException, ImageUploadException {
+    public void preUp() throws TwitterException, MediaUploadException {
         uploadUrl = "http://twitgoo.com/api/uploadAndPost";
         String verifyCredentialsAuthorizationHeader = generateVerifyCredentialsAuthorizationHeader(TWITTER_VERIFY_CREDENTIALS_JSON);
